requirements.txt
pyrogram==2.0.106
tgcrypto
motor==3.4.0
python-dotenv
fastapi
uvicorn
